pyproject.toml
[tool.poetry]
name = "model-optimization"
version = "0.1.0"
description = "A project for model optimization and deployment"
authors = ["Khanh Nguyen <khanhnguyen.qk@gmail.com>"]
readme = "README.md"
license = "LICENSE"

[tool.poetry.dependencies]
python = "^3.12.6"

# Core dependencies - for working with transformers models
transformers = "^4.40.0"
torch = "^2.1.0"
accelerate = "^0.27.0"
# bitsandbytes = "^0.41.0"  # for quantization
numpy = "^1.26.0"
huggingface-hub = "^0.20.0"
python-dotenv = "^1.0.0"
# sentencepiece = "^0.1.99"  # Required for many tokenizers
# safetensors = "^0.4.0"

# Optimization & quantization
# optimum = "^1.16.0"
# onnx = "^1.15.0"
# onnxruntime = "^1.16.0"

# NVIDIA-specific optimizations
# tensorrt = "^8.6.0"
# torch-tensorrt = "^1.4.0"
# optimum-nvidia = "^1.0.0"

# for GPTQ quantization
# auto-gptq = "^0.6.0"

# Inference server - Triton
# triton-model-analyzer = "^1.32.0"
# tritonclient = "^2.41.0"

# Benchmarking
# deepspeed = "^0.12.0"  # performance optimization
# evaluate = "^0.4.0"
# datasets = "^2.16.0"
matplotlib = "^3.8.0"
pandas = "^2.1.0"
seaborn = "^0.13.0"  # Visualization

# Utilities
# tqdm = "^4.66.0"  # Progress bars
# pyyaml = "^6.0.0"  # Configuration files
# jupyter = "^1.0.0"  # For notebook development

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 